chore(timer): split inline tests out of timer.rs after #8328
#8328 pushed timer.rs to 2001 lines, one over the check_file_size.sh cap.
Move the trailing #[cfg(test)] block to timer/tests_inline.rs behind a
#[path] mod, following the box/release_tests.rs precedent, and re-export it
so the helpers other modules reach as crate::timer::* still resolve.
