+++ changelog.d/8355-split-timer-tests.md
@@ -1,0 +1,1 @@
+Split `crates/perry-runtime/src/timer.rs`'s inline unit tests into `timer/tests_inline.rs`: #8328 pushed the file to 2001 lines, one over the 2000-line cap enforced by `scripts/check_file_size.sh`.
